requirements.txt
requests
beautifulsoup4
pandas
numpy
sqlalchemy
python-dotenv
sklearn
pymc
seaborn
matplotlib
plotly